backend/requirements.txt
google-generativeai
langchain-google-genai
langchain-community
chromadb
pypdf
ipython
pandas
langchain
fastapi
uvicorn
load-dotenv
unstructured
python-magic
python-pptx
docx2txt
